Remove all code to make way for the next iteration
This is the start of a rewrite of this package.

After using the package for several differeny projects, by several
different people, we've learned what works, and what doesn't.

Some less than ideal implementations that will be tackled in v2:

  * adding new stream clients that behave slightly differently is
    problematic
  * the shared configuration state between producers and consumers for
    the same client is less than ideal
  * package naming and separation of concerns can be improved
  * the different client implementations don't have a uniform way of
    handling channels and closing them

Instead of doing a step-by-step refactor of the existing code, it proved
to be a much better exercise to remove all existing code, and start from
scratch, only re-using specific code blocks where it makes sense.