Move from interface-based to struct-based message handling
While the experiment with interface-based message handling was an
interesting one to have, and a good learning experience, in the end it
resulted in a less intuitive and user-friendly experience[1].

The new solution is a mix between old and new, and hopefully it hits the sweet
spot.

Specifically, we now have a fixed set of fields that we support for each
message:

* Key string
* Value []byte
* Timestamp time.Time
* Offset *int64
* Tags map[string][]byte

In case of messages returned by a consumer, all these fields are _read only_,
meaning they can be set, or not, and you can use them in whatever way makes
sense.

Most keys return to their default empty values if not set (such as empty key
strings, or a tags map of length 0). The Offset is a pointer, to allow the zero
value to be nil. Any other value than nil means it's the real offset of that
message.

In case of the producer, the producer implementation itself can behave
differently based on the field values provided for a specific message. Some
fields might be ignored by producers, either because they have no meaning for
that specific implementation (e.g. the Timestamp field for the standardstream
client implementation), or because they are only relevant for the consumer and
not for the producer (e.g. the Offset field for the Kafka client
implementation).

The meaning of any of these fields is fairly self-explanatory, but it's still
up to the consumer and producer implementations to provide the real meaning of
each field. For example, in the case of tags, for the Kafka implementation,
these represent the "header" property of a Kafka message, whereas in Pubsub,
they represent the "labels" property, fairly similar in scope, but still
slightly different.

Finally, a message has a semi-private `opaque` field attached to it. This field
is used by client implementations to attach invisible data to a message. This
data can be of any kind (it's an empty interface), and it's up to the client
implementation to use this in whatever way makes sense. As an example, for the
Kafka consumer implementation, this field stores the consumes message's
so-called "topic partition" information. Whenever `consumer.Ack(message)` is
called, this topic partition information is extracted from the message, and used
to actually acknowledge the message at the Kafka brokers.

Speaking of acknowledging messages; another change that made this move towards
struct-based messages possible, was to move the `Ack` method away from the
actual message struct, and onto the consumer struct. This has several
advantages:

* The Ack (and Nack) method requirement is now part of the consumer interface
* The message no longer has to support one giant (N)Ack implementation to
  support all different client implementations.
* It is more clear that (N)Ack is part of the consumer spec, and has no relation
  to the producer.

So now, instead of doing this:

    message.Ack()

You do this:

    consumer.Ack(message)

The only downside to this is that you'll potentially have to make this method
available deep down a stack, requiring you to pass along the consumer. Of
course, one way to partially solve that is to only pass the Ack function itself
as an argument, using the method interface `func(streammsg.Message) error`.

We'll see how all of this pans out, but so far the resulting experience is
better, without sacrificing too much in terms of message bloat, or too
implementation specific message fields.

1: http://idiomaticgo.com/post/best-practice/accept-interfaces-return-structs/