Add standardstream.NewConsumer/Producer functions
These functions are the main entrypoint into the functionality of this
package.

It is fairly straight forward:

In the case of NewConsumer, if you set the `STREAMCLIENT_CONSUMER`
environment variable to the name of one of the streamclient
implementations, it will return that implementation. Valid names are:

- standardstream
- inmem
- kafka
- pubsub*

If you _don't_ pass in the environment variable, but send data to the
application over Stdin, the "standardstream" consumer will be returned
from this function:

    echo '{"hello": "world" }' | go run main.go

In the case of NewProducer, the same applies as the consumer, except
that the environment variable is named `STREAMCLIENT_PRODUCER`.

Also, there is no concept of Stdin for the producer, but if you set
the environment variable `DRY_RUN` to anything other than an empty
string, the "standardstream" producer will be returned, but _only_ if
the `STREAMCLIENT_PRODUCER` environment variable isn't also set.

In the case of both functions, if the right streamclient implementation
cannot be determined by the provided context, none will be chosen, and an
error will be returned instead.

* pubsub has yet to be implemented, and will return an error for the time
being.