Introduce `Errors()` channel for consumers and producers
Every producer and consumer now has its own internal errors channel to
which any errors are delivered that occur during consuming or producing
messages from or to the stream.

By default, this errors channel is consumed internally, and any error
received on the channel is logged on a "fatal" level, indicating the
message is sent to stderr, and the application exits with status code 1.

This default behavior can be changed, by setting the `HandleErrors`
configuration flag to false, indicating that the consumer or producer
should not listen to the errors channel. When configured like this, the
processor application itself is required to listen to any errors being
sent to the `Errors()` channel, and the processor should dictate what to
do with each error.

For now, the errors are delivered in raw form. In the future, this might
be changed to allow the consumer to more easily distinguish between the
different error types returned on the channel.