Move streammsg.Message to stream.Message
This makes it easier to use the Message struct, and it keeps the same
message package name as version 1 of the library.

To prevent a circular import cycle, the `Config()` function now returns
an interface that can be cast to either a `streamconfig.Consumer` or
`streamconfig.Producer` struct.

This is an acceptable trade-off, as you really only use this function
sparsely (and probably only in test scenarios), whereas you use the
`Message` struct constantly.