Add clear separation between ConsumerTopic and ProducerTopic
Before, a consumed message would set `Topic` field to the topic the
message originated from, while the producer would use that same field to
dictate where a message should be produced to.

This was _super_ dangerous, as you can see a processor doing something
like this:

    for msg := range consumer.Messages() {
      // do some stuff with `msg`

      producer.Messages() <- msg
    }

Even if the producer was configured to use a specific topic, it is always
possible to overwrite that configured topic on a per-message basis by
setting the `Topic` field. Since that field is also set by the consumer,
you could end up in a situation where consumed message was produced to the
same topic as from where the message was consumed.

This resulted in a single Message struct field serving two very distinct
purposes, and while the difference was explained in the documentation,
this clearly violates the principle of least surprise.

With this introduced change, there is now a clear distinction between the
`ConsumerTopic` value, and the `ProducerTopic` value. Obviously, the first
is set by the consumer and can be used to identify the topic a messages
originated from, while the latter can be set by the application to
indicate that this message should not be delivered to the pre-configured
producer topic, but to the configured `ProducerTopic` topic.