Add `streamutil.Interrupt()` function
While not _directly_ related to streams or stream processors, interrupt
handling is a _very_ important aspect of stream processing, as you want as
much control over when you stop your processor as possible.

This is why this library has had correct "default" interrupt behavior
from the start (essentially cleanly closing any open consumers or
producers when required).

However, cleanly closing a producer or consumer is not always enough, as
your business logic is very much a driving factor in dictating what the
"correct" behavior is when an interrupt or kill signal is received.

For example, if your business logic reaches out into data stores other
than the streams you are consuming from or producing to, you might want to
wait for those actions to be finished cleanly, before stopping your
application.

You already have control over whether or not use the built-in signal
handling by explicitly disabling it, like so:

    streamclient.NewConsumer(streamconfig.ManualInterruptHandling())

At this point, interrupt handling would be disabled, and you'd have to
manually manage interrupts. It's not like this is hard to do, but this
function makes it a tiny bit easier in the context of a select statement
which loops over incoming stream messages, like so:

    for {
      select {
      case msg := <-c.Messages():
        // handle the incoming message, we'll consider any business logic
        // inside this case statement a a loose transaction, and want to
        // make sure it finishes before we terminate the application.

      case <-streamutil.Interrupt():
        // at this point, we've received an interrupt signal, and we also
        // know that the previous message processing finished, and any
        // future message processing is on hold. We can now cleanly exit.
        return
      }
    }

The above code is thread-safe: if you have multiple goroutines handling
multiple consumed topics, you can call this function in all of those
goroutines, and you'll receive the interrupt signal in each of them for
you to act on as you please.