Add support for `KafkaGroupIDRandom` option
The `streamconfig.KafkaGroupIDRandom()` option allows you to easily assign
a random group ID to the consumer every time you start that consumer.

This is useful for one-off consumers that aren't allowed to share consumer
group state with other consumers, or between runs.

This can be used in combination with `streamconfig.KafkaOffsetHead()` or
`streamconfig.KafkaOffsetTail()` to make sure the consumer starts reading
from that configured offset every time you start the consumer, instead of
picking up where it left off in the previous consumer group ID.

A concrete example would be if you want to always get the last message in
a specific topic. To do this, you would provide the following two options:

    streamconfig.KafkaGroupIDRandom()
    streamconfig.KafkaOffsetTail(1)