Add support for absolute (tailing) default offsets
This allows you to set an `OffsetDefault` for kafka. There are two
different meanings for this value, depending on if you provide a positive
or a negative integer.

A positive integer means "start consuming from this exact offset value".

A negative integer means "start consuming from the end, minus this value".

For easy configuration, these can be set using the following two options:

    streamconfig.KafkaOffsetHead(i uint32)
    streamconfig.KafkaOffsetTail(i uint32)

There are two important things to note here:

First, this configuration is inherently different from the `OffsetInitial`
property supported by Kafka. The initial property is defined as follows:

> Action to take when there is no initial offset in offset store or the
> desired offset is out of range

Meaning that this value will be used whenever no offset is stored at the
Kafka brokers, or when the desired offset is out of range. There are only
two values Kafka supports here: start from the beginning, or start from
the end. For these, we already have `kafkaconfig.OffsetBeginning`, and
`kafkaconfig.OffsetEnd`.

Which brings us to the second point. The only way to set a specific offset
for the consumer to use as a starting point, is to overwrite the offset
assigned by the broker to the consumer. This means that normally, if you'd
set a hard-coded offset (either a direct offset, or a "tailing" offset),
whatever offset is known by the brokers for the provided consumer group
would simply be ignored, and starting the consumer twice, would result in
the consumer starting to read from the exact same offset each time.

To follow the principle of least surprise, that's not how this library
handles these values. In this case, as the name `OffsetDefault` implies,
if the configured consumer group already has an offset stored for the
provided topic at the brokers, it will use _that_ offset, instead of
whatever you pass in as the default offset. If however the brokers have no
known offset stored, it will start at the default configured offset, and
once the first offset commit is sent to Kafka, the next time the consumer
is started with the exact same configuration, it will start from the
offset known at the broker, not the default offset configured.