Support zero-based default offset
Since offsets are zero-based, setting the default offset to 0 is a
valid use-case.

This change makes OffsetDefault a pointer to an int64 value, making it
possible to distinguish between a `0` offset, and an unset (`nil`)
offset.

The convenience options `KafkaOffsetHead` and `KafkaOffsetTail` still
work the same as before, accepting an int32 value.