Swap KafkaGroupIDRandom from rand int64 to uuidV4
Before, `streamconfig.KafkaGroupIDRandom()` would generate a random
group ID based on a random int64 value. It required you to pass in a
seed value to make sure the returned ID was _actually_ random.

Now, no argument is required, and instead the group ID is now based on a
generated UUID V4 string.

There are several reasons for this change:

- the interface is easier (no more arguments required)
- it is in-line with how `kafkacat` handles random group IDs
- it increases the randomness, reducing group ID collisions