Add convenience function to handle stream errors
I came to the realization that the current error handling is flawed in a
sense that we use the provided Zap logger's `Fatal` method to log the
error.

This method then calls `os.Exit(1)`, see also: https://git.io/vhf3n.

The problem with this is that `os.Exit` does _not_ allow any deferred
functions to be executed, as it exits immediately, this unlike `panic`,
which _does_ allow the stack to complete before terminating.

But even if we switch to using `panic` instead, it would still not be
enough, because we could only handle a correct termination of the
_current_ consumer or producer, but if a processor uses multiple
consumers and/or producers, the others would still terminate without
properly cleaning up.

Cleaning up is especially important in the case of a Kafka client, since
it heavily uses internal caches of messages either to be consumer by the
client, or to be produced to the broker. If a producer for example
terminates without draining its queue, it could result in already
acknowledged messages by the consumer not being delivered by the
producer.

To solve this, I think two changes are required:

1. make it easy to handle errors from multiple consumers and/or
   producers by the processor itself, without adding too much overhead.

2. either make manual error handling the default, _or_ make it
   explicitly clear that the default isn't as fault tolerant as the
   manual solution is.