Add helper function to marshal messages in Zap log entries
This change adds the `streamlog` package. It currently exposes a single
function:

    func Message(m stream.Message) zap.Field

You pass in a `stream.Message`, and a `zap.Field` is returned, that can
be directly passed into a Zap log entry:

    logger.Info("My message details.", streamlog.Message(msg))

For simplicity and consistency, the field key cannot be changed. It
will be set to `streamMessage`. If you want to change this field, you
can implement your own marshaller.